refactor(ui): extract usePaneFileDrop hook and dedupe drop wiring
Post-review cleanup: move the per-pane OS file-drop wiring out of the
already-large LeafPanelView into a dedicated usePaneFileDrop hook, fold
the stray FileInput import into the existing lucide-react block, and
guard useOsFileDrop's setState so continuous drag-over events only
re-render a pane when its hover state actually flips.
