+++ src/components/SplitView/SplitView.tsx
@@ -13,6 +13,7 @@ import * as ContextMenu from "@radix-ui/react-context-menu";
 import {
   Settings as SettingsIcon,
   FileEdit,
+  FileInput,
   SquarePen,
   ScrollText,
   ArrowLeftRight,
@@ -32,12 +33,9 @@ import {
 } from "lucide-react";
 import { useAppStore } from "@/store/appStore";
 import { PanelNode, LeafPanel, TerminalTab, DropEdge } from "@/types/terminal";
-import { getAllLeaves, findLeafByTab, getPanelActiveSessionId } from "@/utils/panelTree";
+import { getAllLeaves, findLeafByTab } from "@/utils/panelTree";
 import { isWindows, isMac } from "@/utils/platform";
-import { useOsFileDrop } from "@/hooks/useOsFileDrop";
-import { sendInput } from "@/services/api";
-import { quotePath } from "@/utils/quotePath";
-import { FileInput } from "lucide-react";
+import { usePaneFileDrop } from "@/hooks/usePaneFileDrop";
 import { writeText as writeClipboard } from "@tauri-apps/plugin-clipboard-manager";
 import { ConnectionIcon } from "@/utils/connectionIcons";
 import { useTerminalRegistry } from "@/components/Terminal/TerminalRegistry";
@@ -567,17 +565,7 @@ function LeafPanelView({ panel, setActivePanel, activeDragTab }: LeafPanelViewPr
   // OS file drop is scoped per pane: a path dropped onto this panel is inserted
   // into the session of the tab shown here, so each open connection is its own
   // drop target rather than the whole terminal area being a single zone.
-  const terminalAreaRef = useRef<HTMLDivElement>(null);
-  const panelSessionId = getPanelActiveSessionId(panel);
-  const handleFileDrop = useCallback(
-    async (paths: string[]) => {
-      if (!panelSessionId || paths.length === 0) return;
-      const text = paths.map(quotePath).join(" ");
-      await sendInput(panelSessionId, text);
-    },
-    [panelSessionId]
-  );
-  const { isDragOver: isFileDragOver } = useOsFileDrop(terminalAreaRef, handleFileDrop);
+  const { terminalAreaRef, isFileDragOver, panelSessionId } = usePaneFileDrop(panel);
 
   // Capture selection BEFORE right-click modifies it (xterm auto-selects word on right-click)
   const preRightClickSelectionRef = useRef<string | null>(null);
